# Cache config JSON to speed up extension config access