--- a/edb/buildmeta.py
+++ b/edb/buildmeta.py
@@ -44,7 +44,7 @@ from edb.common import verutils
 
 
 # Increment this whenever the database layout or stdlib changes.
-EDGEDB_CATALOG_VERSION = 2023_10_02_00_02
+EDGEDB_CATALOG_VERSION = 2023_10_02_00_03
 EDGEDB_MAJOR_VERSION = 4