--- a/edb/server/pgcon/pgcon.pyx
+++ b/edb/server/pgcon/pgcon.pyx
@@ -122,13 +122,19 @@ cdef object logger = logging.getLogger('edb.server')
 # * 'A': an instance-level config setting from command-line arguments
 # * 'E': an instance-level config setting from environment variable
 SETUP_TEMP_TABLE_SCRIPT = '''
-        CREATE TEMPORARY TABLE _edgecon_state (
-            name text NOT NULL,
-            value jsonb NOT NULL,
-            type text NOT NULL CHECK(
-                type = 'C' OR type = 'B' OR type = 'A' OR type = 'E'),
-            UNIQUE(name, type)
-        );
+    CREATE TEMPORARY TABLE _edgecon_state (
+        name text NOT NULL,
+        value jsonb NOT NULL,
+        type text NOT NULL CHECK(
+            type = 'C' OR type = 'B' OR type = 'A' OR type = 'E'),
+        UNIQUE(name, type)
+    );
+'''
+SETUP_CONFIG_CACHE_SCRIPT = '''
+    CREATE TEMPORARY TABLE _config_cache (
+        source edgedb._sys_config_source_t,
+        value edgedb._sys_config_val_t NOT NULL
+    );
 '''
 
 def _build_init_con_script(*, check_pg_is_in_recovery: bool) -> bytes:
@@ -149,10 +155,14 @@ def _build_init_con_script(*, check_pg_is_in_recovery: bool) -> bytes:
         {pg_is_in_recovery}
 
         {SETUP_TEMP_TABLE_SCRIPT}
+        {SETUP_CONFIG_CACHE_SCRIPT}
 
         {INIT_CON_SCRIPT_DATA}
 
         PREPARE _clear_state AS
+            WITH x1 AS (
+                DELETE FROM _config_cache
+            )
             DELETE FROM _edgecon_state WHERE type = 'C' OR type = 'B';
 
         PREPARE _apply_state(jsonb) AS
@@ -765,6 +775,9 @@ cdef class PGConnection:
     def _build_apply_state_req(self, bytes serstate, WriteBuffer out):
         cdef:
             WriteBuffer buf
+
+        if self.debug:
+            self.debug_print("Syncing state: ", serstate)
 
         buf = WriteBuffer.new_message(b'B')
         buf.write_bytestring(b'')  # portal name