Every config object and multi link view makes a call to
_read_sys_config, which adds up pretty badly when accessing extension
configs.

Add a temporary _config_cache table that is used to cache results, and
modify _read_sys_config_full to be able to compute the output for all
three relevant config levels at once.

`select cfg::Config.extensions[is ext::_conf::Config].config_name`
goes from taking about 50ms on average to taking ~23ms on a cache
miss and ~6.5ms on a cache hit.
(Before I removed instance level extension config objects in #6193,
it took ~300ms.)

Probably the last part of #5954 that will get done before 4.x.