@msullivan 2023-10-03
I had planned to specifically write a test that verified that the `__dbver__` cache entry was working properly (that would mimic how I tested it manually), but when I disable the `__dbver__` cache entry, a ton of tests fail, and I am now convinced it doesn't need a new test to exercise it